requirements.txt
aiogram>=3.4.1
google-genai>=0.5.0
python-docx>=1.1.0
fpdf2>=2.7.9
python-dotenv>=1.0.1
Pillow>=10.2.0
PyPDF2>=3.0.0
aiohttp>=3.9.0
